Autocomplete: use a smaller percentage to detect the same string (#1893)
- While working on dynamic multiline completion, I extensively tested
autocomplete manually and noticed that helpful completions were hidden
in multiple cases. It happened because `isAlmostTheSameString` decided
that the generated completion matches the string after it. This PR
lowers the percentage defined in `isAlmostTheSameString` to fix the
issue.
- The main reason for this functionality was the lack of
fill-in-the-middle support by the existing models, so LLMs generated
completion, duplicating the current code. That's not the case anymore,
and we may not need this logic or can lower the percentage further.