Agent: use `uri: vscode.Uri` instead of `filePath: string` (#1928)
Previously, the agent used a property named `filePath` as the unique
string identifier of documents. This was problematic for several
reasons, including

- There's no idiomatic way to encode virtual files
- File paths have different formatting in Windows and Linux/macOS

This PR does a big refactoring around the agent internals so that we use
URIs instead of `filePath`. URIs solve both problems above.

The motivation to prioritize this change was that bfg wasn't loading
correctly in the evaluation because of a bug related to handling of
relative file paths. This bug was trivial to prevent after moving to
URIs.

The diff includes a few minor `evaluate-autocomplete`-related changes
that I included since I wanted to validate that bfg evaluation works
correctly after the refactoring.

## Test plan

Green CI.
<!-- Required. See
https://docs.sourcegraph.com/dev/background-information/testing_principles.
-->