Autocomplete: dynamically switch to multliline completions (#1894)
- We now use the completion's first line to determine if it
starts the new multiline syntax node. If it does, we continue streaming
until the completion is truncated or we reach the token sample limit.
- Closes https://github.com/sourcegraph/cody/issues/1619
- Adds a **feature flag**
`cody-autocomplete-dynamic-multline-completions` to run an experiment on
the updated logic.
- Adds an **experimental setting**,
`cody.autocomplete.experimental.dynamicMultilineCompletions` to test the
feature locally.
- Adds `completionPostProcessLogger` used for local debugging. It's
disabled by default, and I will replace it with OpenTelemetry spans in a
follow-up PR. I put a little effort into improving this logger because
it's a throw-away code, but I need it here because debugging
streaming-truncation bugs is much faster with it.
- Revamps the multiline truncation logic:
- We no longer look for the next new sibling after the completion. This
approach is unreliable because the parse tree often breaks because of
incomplete code in the completion insert text.
- Instead, we take the descendant at the current cursor position (or
multiline trigger position), look for the farthest ancestor that starts
at the same line, and then wait for a moment when completion generates
more line than this syntax node has. This means we do not care about
broken parse-tree after the node under the cursor.