Add rate limit notification/upgrade buttons in chat (#1882)
I've updated the error handling so that `RateLimitError` (instead of
just a string) is propagated up to the chat. With a fake rate limit
(using the same status code + headers as the existing completion rate
limit), it looks like this:


![image](https://github.com/sourcegraph/cody/assets/1078012/eb5f1ae7-cb32-44e1-8abb-3440cfe9b789)

## Test Plan

- There are some e2e tests that call fake methods on the mock server to
set rate limit status and verify the appropriate messages show up
- This will require manual testing against a real service to verify the
code/tests match what the live servers are doing. In particular:
  - non-dotCom should never show upgrade messages
  - dotCom should show upgrade messages if not already Pro
  - dotCom should not show upgrade messages if already Pro