Agent: add parsing test to evaluate-autocomplete
Previously, evaluate-autocomplete only supported running a "typecheck"
test that was 1) slow 2) difficult to configure and 3) could have false
results for cases where the build command didn't typecheck all files.
This PR adds a sibling "parse" test (enabled by default) that is super
fast to run and requires no configuration.

This PR also adds multiple new columns to the generated CSV to make it
easier to run analysis on the data.