Fix flaky e2e chat tests (#2061)
This fixes two flakes in recent tests I added:

1. VS Code toast notifications obscuring the chatbox so the test could
not click them.

The failure looked like:

```
 retrying click action, attempt #2
      waiting 20ms
      waiting for element to be visible, enabled and stable
      element is visible, enabled and stable
      scrolling into view if needed
      done scrolling
      <div title="Git (Extension)" class="notification-list…>Source: Git (Extension)</div> from <div class="notifications-toasts visible">…</div> subtree intercepts pointer events
```

This was fixed by turning on Do Not Disturb mode (something already done
in another test, so I extracted a function and reused it).

2. Triggering rate limits in the second test sometimes not working.

This one was confusing. It appears that in e2e tests we stop the server,
we just call `server.close()` which only prevents new connections. The
server continues to run for any open connections. This could result in
the next test from reusing some existing connections to the first server
and creating some new connections to its own new server. For tests that
modify state in the mock server (eg. to trigger rate limits) this would
cause failures.

Adding unique IDs to the server and some logging resulted in this:


![flakytest](https://github.com/sourcegraph/cody/assets/1078012/2dd17c8d-ea23-4288-9434-27c8c2ec0256)

The fix is to ensure all existing connections to the server are closed
when we close the server (by collecting them as they open and calling
`destroy()` on them).


## Test plan

These are fixes to the test, green bots with no failures/retries
indicate it works.