vscode: Remove App login (#2099)
This one is a doozy:

- Removes App as a sign-in option.
- If you're signed into App but you have a dotcom token, we'll try to
use that.
- If you're signed into App, but don't have a dotcom token, you'll have
to sign in again.
- Even if you're using classic sidebar chat, if you had an embeddings
index generated by App, we'll pick it up and use it. However the context
status indicator will indicate you have no embeddings. (We could plumb
this in, but it is going away very soon.)
- No more toasts or promotions to set up App.

Leaves the LocalAppDetector so we can delete App/toast you to delete App
soon.

Part of #1966.

## Test plan

**Setup**

The big secret to testing this is recreating the "signed in to App"
state, now that all the sign in options are gone. You do this by abusing
the Enterprise sign-in options:

1. Install Cody App and log into it if you haven't got it already.
2. Sign into dotcom, which caches a token used for the first test below.
3. In VScode, Cmd-Shift-P, Cody: Switch Account... and enter
`http://localhost:3080` which is App's URL.
4. cat ~/Library/Application\ Support/com.sourcegraph.cody/app.json and
smurf the token.
5. At this point, you should be signed in to App. You can verify this by
chatting, quitting app, chatting again and seeing you get a connection
error message.

**Next sign in**

1. Simply reload the window.
2. You should be signed in to dotcom with the token you had been using
at the start of the previous step. You can verify this by chatting with
App closed.
3. Check that Cody App doesn't appear in any of the account switcher or
sign in menus.

**Signed into App, no dotcom token**

Assuming you're still signed in to dotcom:

1. Cmd Shift P, Cody: Sign out.
4. Sign into App using the leetcode techniques above.
5. Reload the window.

You should be dumped at the login pane.

**Signed into App, cached invalid dotcom token**

1. Create a new dotcom auth token at
https://sourcegraph.com/user/settings/tokens/new/callback
2. Sign in to http://sourcegraph.com using that token
3. Cmd Shift P, Cody: Switch Account... and follow the "setup"
instructions above to sign in to app.
4. You are now signed in to App, with a *valid* cached dotcom token.
Switch back to the Sourcegraph token management UI and revoke the token.
5. Reload the window. This encounters a *long* (60s?) pause, but
eventually the login fails and you are dumped back to the login screen
and can log in to get a fresh token.