evaluate-autocomplete: more refined triggering
This PR changes the locations where we trigger requsts to get a bigger
breadth of examples. Previously, we could use up the request limit on
similar code patterns that repeated on multiple lines. A lot of the
requsts were also spent on locations where the result was trivial. This
PR changes that by:

- Adding a limit on the minimum size of the match so that we only
  trigger autocomplete in locations where we expect a large autocomplete
  suggestion
- Adding an option to trigger only in every N-th match.

Previously, we only triggered autocomplete in limited cases. This PR adds new pluggable infrastructure to trigger in a more variety of language constructs, specifically call expressions, function declarations, if statements, and assignment statements. This ability allows us to measure more accurately real-world usage behaviors.

-- Test plan

I ran an evaluation of 8 languages to compare scores between baseline and dynamic-multiline. The score is a formula that penalizes autocomplete requests that don't parse or are very small.

```
             FIXTURE  LANGUAGEID      SCORE
0           baseline         cpp  64.033333
1           baseline          go  28.811475
2           baseline  javascript  40.990099
3           baseline         php  39.079365
4           baseline  typescript  46.227106
5  dynamic-multiline         cpp  83.815789
6  dynamic-multiline          go  74.201681
7  dynamic-multiline  javascript  76.683673
8  dynamic-multiline         php  58.024194
9  dynamic-multiline  typescript  78.115942
```

When plotting these scores on a radar chart we see that dynamic-multiline is a huge improvement across the board including languages like PHP.