evaluate-autocomplete: minor tweaks
This PR changes the locations where we trigger requsts to get a bigger
breadth of examples. Previously, we could use up the request limit on
similar code patterns that repeated on multiple lines. A lot of the
requsts were also spent on locations where the result was trivial. This
PR changes that by:

- Adding a limit on the minimum size of the match so that we only
  trigger autocomplete in locations where we expect a large autocomplete
  suggestion
- Adding an option to trigger only in every N-th match.