Autocomplete: Add Experimental hot streak mode (#2118)
Closes #2205

This PR adds a new feature flag and experimental config option to enable
_hot streak mode_.

The idea is simple: When generating a completion (and this is mostly for
single-line and works beautifully with dynamic-multilin), we let the LLM
continue to generate more than just the current line and use follow-up
lines to seed a cache.

Then, when a user accepts a completion and moves to the next line by
pressing enter, we can instantly generate another completion (and thus
avoid doing another LLM request and incurring latency). The result is a
UX where monotonous, single-line completions appear a lot faster and
allows you to fall into a tab+enter, tab+enter, … mode in which you can
review a longer completion line by line. It’s also going to be much
faster than multiline-completions because we can show the first line as
soon as it is ready.

⚠️ This is currently limited by #2180

## Test plan

- Enable hot streak mode
- Observe that after accepting a completion, pressing enter will show
the continuation of that completion ✨ instantly ✨

https://github.com/sourcegraph/cody/assets/458591/5f70ab45-18a2-4a4c-8d02-61a45328df47