Edit: Do not scroll into view on complete (#2297)
closes https://github.com/sourcegraph/cody/issues/2259

## Description

Previously we would always scroll a completed task into view on
completion.

We have a bug with this already
(https://github.com/sourcegraph/cody/issues/2259).

One fix here is just to scroll this into view when the code is NOT
already in view.

However, I think this might actually be annoying behaviour. The user
already gets a notification for edits that are made if the file is not
currently active. It seems overly aggressive to force their scroll back
to the code when editing an active file.

## Test plan

1. Edit code in a large file
2. Scroll the code out of view
3. Check we don't hijack the scroll


<!-- Required. See
https://docs.sourcegraph.com/dev/background-information/testing_principles.
-->