Agent: add test case to accept autocomplete
Previously, the agent tests didn't test the case where the client sends
a notification for an accepted autocomplete request. Turns out that this
notification caused uncaught errors that may have been the root cause
behind sourcegraph/jetbrains#191

This PR adds a test case to reproduce the situation where a client
accepts an autocomplete request and fixes the errors that happen when
doing so.