Agent: implement `vscode.workspace.fs` (#1716)
Previously, the agent didn't support the `vscode.workspace.fs` API. This
API is used in various places in the Cody extension to read/write files
to disk. The lack of `vscode.workspace.fs` support for non-vscode
clients represented an important feature gap. This PR fixes the problem
by implementing this part of the VS Code API so that clients like
JetBrains and Neovim can benefit from Cody features that interact with
the file system.

## Test plan

Unit tests added in
[vscode-shim.test.ts](https://github.com/sourcegraph/cody/blob/main/agent/src/vscode-shim.test.ts)

---------

Co-authored-by: Olafur Geirsson <olafurpg@gmail.com>