HACK: Work around rate limiting errors by parsing the SSE message and reverse-engineering the headers (#2375)
This attempts to work around the rate limiting header issues and instead
parses the first SSE payload error text and tries to reverse-engineer
the PLG headers.

To do that, we parse the error string, e.g.:

```
Sourcegraph Cody Gateway: unexpected status code 429: you have exceeded the rate limit of 10 requests. Retry after 2023-12-15 14:36:37 +0000 UTC\n
```

To get the quota and the retry after field, and make a request to the
backend to find out if the user is already a pro user.

https://github.com/sourcegraph/cody/assets/458591/fca610e3-f3a5-4b5d-9f0c-5b765e18512f

https://github.com/sourcegraph/cody/assets/458591/ebba8005-0ee9-4ee8-825d-637415252689

---------

Co-authored-by: Beyang Liu <beyang.liu@gmail.com>