Agent: add support for new VS Code chat UI
Previously, the agent only supported chat via recipes, which had an
increasingly large feature gap with chat in VS Code. This PR extends
the agent protocol with new high-level endpoints to support interacting
with the enw chat UI:

- `chat/new`: requests to start a new chat. Returns an ID number to
  reference this chat session.
- `chat/submitMessage`: request to submit a chat message. The agent
  streams progress via a separate `webview/postMessage` notification,
  and responds to this request when the last message has finished
  streaming.

The endpoints above are "high-level", meaning they are implemented on
top of more low-level endpoints that directly translate VS Code API
calls:

- `webview/create`: request from the server to the client to create a
  new webview. Wrapper for `vscode.window.createWebviewPanel`.
- `webview/postMessage`: notification from the server to the client to
  send a JSON message to the webview.
- `webview/receiveMessage`: notification from the client to the server to
  send a JSON message from the webview.

This design allows us to both support a semi-dropin replacement of the
recipe API while additionally allowing clients to fully interact with
the low-level webview API. In the short term, the JetBrains plugin can
quickly migrate to the high-level APIs. In the long-term, it probably
makes sense for JetBrains to use webviews and use the low-level API.