Autocomplete: Do not trigger a completion if a single-line completion was accepted (#2446)
This is a small optimization that should cut down a bit on unnecessary
LLM request:

In VS Code, if you accept a completion, this changes the document and
thus issue a new completion request. Makes sense! The thing is: we
always complete full lines so the chance that you do want a follow-up
completion is _very_ slim. In some cases the ending `)` or `;` helps to
guard against this but not in all cases.

This PR adds a special logic now to detect this case and not issue a
completion (so accepting a completion via `tab` won't issue a new
request but any further keystroke (like `enter`) will).

## Test plan

- Added tests
- 


https://github.com/sourcegraph/cody/assets/458591/83d9c044-a886-4e72-90bf-d2c1486f260e



<!-- Required. See
https://sourcegraph.com/docs/dev/background-information/testing_principles.
-->

---------

Co-authored-by: Dominic Cooney <dominic.cooney@sourcegraph.com>