Remove serverEndpoint from config object (#2456)
This refactors the `Config` object to no longer expose a
`serverEndpoint`. The `serverEndpoint` should be retrieved only from
local storage as part of the authenticated user. When going through this
legacy code part, the server endpoint would always be set to the dotcom
instance URL instead.

This caused some parts to behave like you were on dotcom even when
inside an enterprise version.

## Test plan

- I tested it by evaluating analytics logs and switching accounts
- Video is large so uploading it on Google Drive instead:
https://drive.google.com/file/d/1dpn-C1IczJ9CQ0zX6evMyzKpHpq8lIvS/view?usp=sharing