Chat: Scope history to user account (#2348)
Closes #2261

This PR scopes the chat history storage to the user account (we use the
email + server URL as the key for this). That way, when switching
accounts, you will not see chat histories from other accounts
(potentially showing work stuff in your free time).

There's one caveat that I have depicted in the test plan. If you keep a
window open while you switch accounts and _send a new message to it_, we
will _fork_ the history, so it will appear in your other account with
the updated messages but kept as-is in your other account. I think this
behavior makes sense though and we should not try to optimize for it
(i.e I don't think we should close chat windows, for example, when you
switch accounts).

⚠️ I plan to land this after cutting the 1.0.2 patch release and deploy
it onto the insider build right away to test this some more.

## Test plan

- Log in to a user from dotcom
- Chat!
- Switch to another account, e.g. S2
- More chatting!
- ...repeat

https://github.com/sourcegraph/cody/assets/458591/a42f1a4b-da0c-4c4e-839a-c3717d4c8bbb