Fix duplication of query when selecting a context file that already exists
The code here appeared to try to preserve suffixes after the filename, but we appear to only support selecting files at the end of the input anyway. By assuming the index of "fileDisplayText" in the input was the one that the user expected to replace, an input like:

```
@foo.txt @fo
```

Would assume the first `@foo.txt` is being replaced, and retain `@fo` (actually duplicating it).

Removing this suffix fixes the issue and as far as I can tell does not break any other functionality. The issue described in https://github.com/sourcegraph/cody/issues/1973 still appears to work as expected with this change.