Pass URIs for some context snippets (#2520)
Previously, we used the relative file name to indicate the source file
of all context snippets. This causes problems, because:
* The user may have more than one workspace folder open, in which case
we don't know how to resolve a relative file name, because we lose track
of which workspace folder it came from
* Context from remote sources (e.g., remote embeddings) isn't attached
to a local file, and we may need to treat these differently when trying
to open the file

Add a URI field to keep better track of the true location of a context
item. Adds creator functions for 2 URI schemes: `cody-file-relative://`
and `cody-embeddings://`.

Also add an `openUri` function that handles opening these URIs and is
more forgiving for `file://` URIs (will attempt to match a relative file
URI with an open workspace folder).

## Test plan

1. Open at least 2 workspace folders
2. Open a file from the 2nd workspace folder
3. Create a new chat and ask a question
4. Attempt to open the context files—these should open properly, whereas
before, nothing occurred

See screen recordings for examples.

Before:
https://github.com/sourcegraph/cody/assets/1646931/4ae6b2a1-ce00-4d3a-86bd-25effae6495c
After:
https://github.com/sourcegraph/cody/assets/1646931/2d72e149-0f0f-4704-9b2b-3fab699a5b3e