Chat attribution UI (#2521)
Part of
[sourcegraph#58993.](https://github.com/sourcegraph/sourcegraph/issues/58993)

Implements a Chat UI for guardrails following direction from
@toolmantim. Let's look at some screenshots:

First with experimental setting turned off:


![off](https://github.com/sourcegraph/cody/assets/153410/7c31707a-74a0-40d8-98cf-e9c32ba1ba7c)

Once we turn on the experimental setting, each snippet in chat gets a
gray shield icon on the right hand side of the pane below snippet. This
is search running:


![running](https://github.com/sourcegraph/cody/assets/153410/37726258-0614-4c94-9126-3a7ab1c4a50c)

The happy path is that we find no attribution. In that case icon turns
green and we update relevant tooltip.


![notFound](https://github.com/sourcegraph/cody/assets/153410/694f9866-e127-48ef-987b-d7534d279978)

Alternative is that attribution is found. In such case icon turns red.
The tooltip lists one repository where attribution was found, and counts
the other hits:


![found](https://github.com/sourcegraph/cody/assets/153410/7135a291-20ce-4fe5-9dff-f2d56dd101fb)

An exceptional case is when guardrails back end is unavailable. In this
case the icon turns yellow and the tooltip updates:


![unavailable](https://github.com/sourcegraph/cody/assets/153410/328233d6-5666-4be0-befc-14d043c1b94d)

## Test plan

Visual and manual.

I'd love some review feedback on how could we go and implement a story
here. I'm not familliar yet with VS Code ui testing.