Revert "refactor: move commands away from recipe (#2542)" (#2575)
The agent tests started failing after merging the PR #2542 when you
re-record the HTTP requests. The tests were passing in replay mode for
some reason, and we suspect it's due to a buggy merge in
`recording.har`. When you `rm -rf agent/recordings` and record from
scratch then the tests fail.

To prevent this from happening again, we're planning to provide scripts
to cleanly update the recordings.

## Test plan

Green CI.
<!-- Required. See
https://sourcegraph.com/docs/dev/background-information/testing_principles.
-->