Autocomplete: adjust leading whitespace in the last candidate text on indentation (#2560)
- It was a tricky one to pinpoint because of the time-sensitive nature
of the bug. It happens when a user initializes a completion generation
by placing a cursor on the new line and then almost instantly tabs 2+
times. For example:

    ```ts
      function test() {
    █
      }
    ```
    
The initial completion generation request comes back in the background
with the insert-text: `\t\tconst x = 1`, but a user already added three
tabs manually:
    
    ```ts
      function test() {
          █
      }
    ```
    
We reuse (`reuseLastCandidate`) the initial completion that was
generated in the background because only whitespace on the current line
was changed, and the previous logic returned the final insert text as
the following:
`lastTriggerCurrentLinePrefixInDocument.slice(currentLinePrefix.length)
+ item.insertText` which didn't account for leading whitespace that user
added manually. Notice extra whitespace in the ghost text:

    ```ts
      function test() {
          █    const x = 1
      }
    ```

This PR fixes the issue by removing the manually added whitespace from
the insert-text if it's present there. Note that the
`isTypingAsSuggested` clause missed this issue because a user adds more
whitespace than there's present in the current line:
`lastCompletion.startsWith(currentLinePrefix) === false`.