Autocomplete: Remove unused feature flags (#2607)
Removing some more feature flags that already concluded or were never
run (= and won't run in this config).

This leaves the implementation inside though as it is still accessible
via overwriting local settings. That's useful in case we want to improve
the existing implementation. We can always bring back the flags if we
need that.

## Test plan

Trivial change; Relying on CI here

<!-- Required. See
https://sourcegraph.com/docs/dev/background-information/testing_principles.
-->