Autocomplete: split up completion item provider (#2572)
- Closes https://github.com/sourcegraph/cody/issues/2570 by moving the
telemetry helper call to the early return, where we check for the number
of visible completions.
- Simplifies the completions processing logic in
`provideInlineCompletionItems()`. We return early if no visible
completions are generated, so we can simplify the following logic
knowing the completions are available and visible.
- Moves the conversion 
- Extracts the `AutocompleteItem` class and `suggestedCompletionItemIDs`
cache into a separate file. Update the interfaces of these entities for
better readability.
- Extracts the `isCompletionVisible` function and related helpers into a
separate file.
- Extracts `FirstCompletionDecorationHandler` into a separate file.