annotate type-only exports/imports with `type`
This fixed an issue with Storybook complaining about not finding type imports (probably because it uses some magic in its bundling). It's also a good hint to the dev about which imports actually affect the output bundle and which are just for `tsc`.

Rerun `pnpm run lint:js --fix && pnpm run format` to recreate this diff.