Autocomplete: Add a new 16b exclusive experiment and document existing flags (#2632)
For
https://docs.google.com/document/d/1P-Hb8BTbz-vMNFXw3gbAvphArsNvS1_8u4ID-phn8gI/edit,
we are considereing moving all traffic to the 16b model. When we first
run an experiment for this, the single-line CAR was better with the
faster model. Part of this might be due to latency (which will improve
with the ST deployment) but parts of this might also just no longer be
relevant since we have made so many changes.

This PR adds a new client side feature flag that we can run in the short
term to gather data on 16b exclusive setups. It will only have an affect
if the user is already inside the `CodyAutocompleteStarCoderHybrid`
experiment, which is currently our main StarCoder enablement flag.

## Test plan

- Be inside `CodyAutocompleteStarCoderHybrid`
- Be inside `CodyAutocompleteStarCoder16B`
- Observe that single-line completions use the 16b model identifier: 
<img width="904" alt="Screenshot 2024-01-09 at 11 46 51"
src="https://github.com/sourcegraph/cody/assets/458591/801682a5-6626-4767-9898-b9fe9cd3be27">


<!-- Required. See
https://sourcegraph.com/docs/dev/background-information/testing_principles.
-->